+++ Beat/User Interface/Preferences/BeatPreferencesPanel.m
@@ -156,3 +156,11 @@
 }
 
 @end
+/*
+ 
+ the garden's heart has swollen
+ under the sun,
+ and the garden mind is slowly
+ being emptied of green memories.
+ 
+ */