+++ Beat/User Interface/Sidebar/BeatSegmentedCell.m
@@ -43,3 +43,11 @@
 }
 
 @end
+/*
+ 
+ when I trust you we can do it with the lights on
+ when I trust you we can do it with the lights on
+ when I trust you we'll make love until the morning
+ let me tell you all my secrets, and I whisper 'til the day's done
+ 
+ */